Introduced withSmartTags logging functionality and migrated TaggedLogger to use it
